[anhalyze_plot_utils] Code cleanup: moving deprecated functions to depr modules.
